requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.1.0
datasketch==1.5.8
Django==4.0.5
django-braces==1.15.0
django-filter==22.1
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
idna==3.4
importlib-metadata==4.12.0
Markdown==3.4.1
mysqlclient==2.1.1
numpy==1.24.1
opencv-python==4.7.0.72
pandas==1.5.2
phonenumberslite==8.13.6
Pillow==9.4.0
pipreqs==0.4.13
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.2.1
pyzbar==0.1.9
requests==2.28.2
scipy==1.10.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.15
yarg==0.1.9
zipp==3.8.1
